+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- alias: add ability to use a different netowrk interface for alias network type (#80)
+
 ### Fixed
 - prepare: fix multiple export port support
 - rc script: extend PATH to make potnet accessible